+++ Assets/Scripts/Chest_Controller.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3a9c5e899805fd0439031b4f141910ff
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
