+++ Assets/Scripts/Player_Controller.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 27d42560d94f83c4fadf8bb0f49ba6b0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
