--- Assets/Test Scene.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: cb0256e4e2024f64a9a6a169249e5389
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
