+++ Assets/Unity.VisualScripting.Generated.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a01e7e8e49b20d448b029db0666ce8c8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
