+++ Assets/Unity.VisualScripting.Generated/VisualScripting.Flow.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4f998fe4082804b4e98d1896b951c946
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
