+++ Assets/Unity.VisualScripting.Generated/VisualScripting.Flow/UnitOptions.db.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4dc9fbe28149d91499cbf9887a3a511e
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
